test(clar2wasm): reorganize property tests into test folder
