+++ clar2wasm/tests/standard/main.rs
@@ -1,0 +1,2 @@
+mod property_tests;
+mod utils;
